% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpheus.R
\name{morpheus}
\alias{morpheus}
\title{Morpheus Heatmap widget}
\source{

}
\usage{
morpheus(x, labRow = rownames(x), labCol = colnames(x), Rowv = NULL,
  Colv = if (symm) "Rowv" else NULL, distfun = dist, hclustfun = hclust,
  reorderfun = function(d, w) reorder(d, w), rowAnnotations = NULL,
  columnAnnotations = NULL, symm = FALSE, na.rm = TRUE, width = NULL,
  height = NULL, ...)
}
\arguments{
\item{x}{numeric matrix of the values to be plotted.}

\item{labRow}{character vectors with row labels to use (from top to bottom); default to rownames(x).}

\item{labCol}{character vectors with column labels to use (from left to right); default to colnames(x).}

\item{Rowv}{determines if and how the row dendrogram should be computed and reordered. Either a dendrogram or a vector of values used to reorder the row dendrogram or NA to suppress any row dendrogram (and reordering) or by default, NULL, see ‘Details’ below.}

\item{Colv}{determines if and how the column dendrogram should be reordered. Has the same options as the Rowv argument above and additionally when x is a square matrix, Colv = "Rowv" means that columns should be treated identically to the rows (and so if there is to be no row dendrogram there will not be a column one either).}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns. Defaults to dist.}

\item{hclustfun}{function used to compute the hierarchical clustering when Rowv or Colv are not dendrograms. Defaults to hclust. Should take as argument a result of distfun and return an object to which as.dendrogram can be applied.}

\item{reorderfun}{function(d, w) of dendrogram and weights for reordering the row and column dendrograms. The default uses reorder.dendrogram.}

\item{rowAnnotations}{Data frame of additional row annotations in same order as x (optional)}

\item{columnAnnotations}{Data frame of additional column annotations in same order as x (optional)}

\item{...}{Additional morpheus options as documented at https://clue.io/morpheus/configuration.html}

\item{colorScheme}{List of scalingMode ("fixed" or "relative"), stepped (Whether color scheme is continuous (FALSE) or discrete (TRUE)), values (list of numbers corresponding to colors), colors (list of colors)}

\item{rowSize}{Heat map column size in pixels or "fit" to fit heat map to current height (optional, defaults to 13)}

\item{columnSize}{Heat map column size in pixels or "fit" to fit heat map to current width (optional, defaults to 13)}

\item{drawGrid}{Whether to draw heat map grid (optional, defaults to \code{TRUE})}

\item{gridColor}{Heat map grid color (optional, defaults to "#808080")}

\item{gridThickness}{Heat map grid thickness (optional, defaults to 0.1)}

\item{drawValues}{Whether to draw values in the heat map (optional, defaults to \code{FALSE})}
}
\description{
Creates a morpheus.js-based heatmap widget.
}
\examples{
library(morpheus)
rowAnnotations = data.frame(1:32)
morpheus(mtcars, rowAnnotations=rowAnnotations, colorScheme=list(values=list(0, 4), colors=list('green', 'black')))

}
\seealso{
\link{heatmap}
}
